% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colored_ice.R
\name{colored_ice}
\alias{colored_ice}
\title{Plot colored ICE curves}
\usage{
colored_ice(
  pred,
  features,
  covar,
  xlabel = features,
  ylabel = "",
  center = FALSE,
  limits = c(NA, NA),
  nCol = NA,
  colors = NULL,
  alpha = 1,
  legend_title = covar,
  legend_position = "right"
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{features}{A character vector containing the names of the features for which the plot should be created}

\item{covar}{A character string indicating the covariate after which the ICE curves should be colored}

\item{xlabel}{An optional character string of the same length as the number of features indicating the x-axis label of the single plots}

\item{ylabel}{An optional character string indicating y-axis label (same for all panels)}

\item{center}{Logical indicating whether ICE curves should be centered at the minimum. Default is FALSE}

\item{limits}{An optional two-entry vector indicating the limits of the y-axis}

\item{nCol}{An optional numeric entry indicating the number of columns when plots are created for several features}

\item{colors}{An optional vector with entries for each level of 'covar'}

\item{alpha}{An optional numeric entry indicating the alpha-level of ICE curves}

\item{legend_title}{A character indicating the legend title. Default is the name of 'covar'}

\item{legend_position}{Logical indicating whether the legend should be shown. Default is TRUE}
}
\value{
a plot of type ggplotify
}
\description{
This function allows to plot colored ice predictions for a set of features, colored by a covariate
}
\examples{
\dontrun{
N <- 1000
x1 <- runif(N, -1, 1)
x2 <- runif(N, -1, 1)
y <- 5 + 5 * x1 * x2 + rnorm(N,1)
dat <- data.frame(x1,x2,y)
rfmod <- randomForest::randomForest(y~., dat)
pred <- iml::Predictor$new(rfmod)
colored_ice(pred, features = "x1", covar = "x2")
colored_ice(pred, features = "x1", covar = "x2", center = TRUE)
}

}
