% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colored_pdp.R
\name{colored_pdp}
\alias{colored_pdp}
\title{Plot colored PD curves}
\usage{
colored_pdp(
  pred,
  feature,
  covar,
  title = "",
  xlabel = feature,
  legend_title = covar,
  legend_position = "right"
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{feature}{A character vector containing the names of the feature for which the plot should be created}

\item{covar}{A character string indicating the covariate after which the ICE curves should be colored}

\item{title}{An optional character string indicating the title of the plot}

\item{xlabel}{An optional character string of the same length as the number of feature indicating the x-axis label of the single plots}

\item{legend_title}{A character indicating the legend title. Default is the name of 'covar'}

\item{legend_position}{Logical indicating whether the legend should be shown. Default is TRUE}
}
\value{
a plot of type ggplotify
}
\description{
This function allows to plot colored pd predictions for one feature, colored by a covariate
}
