% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretabilityMethods-package.R
\docType{package}
\name{InterpretabilityMethods-package}
\alias{InterpretabilityMethods-package}
\alias{InterpretabilityMethods}
\title{InterpretabilityMethods: Facilitates the implementation of interpretability methods for social science research}
\description{
Facilitates the implementation of interpretability methods for social science researchers. The packages uses functions for interpretability methods from the packages 'iml' and 'permimp'.
}
\author{
Mirka Henninger
}
