% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretableML-package.R
\docType{package}
\name{InterpretableML-package}
\alias{InterpretableML-package}
\alias{InterpretableML}
\title{InterpretableML: Facilitates the implementation of interpretation techniques for social science research}
\description{
Facilitates the implementation of interpretation techniques for social science researchers. The packages uses functions for interpretability methods from the packages 'iml' and 'permimp'.
}
\author{
Mirka Henninger
}
