% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretionMethods-package.R
\docType{package}
\name{InterpretationMethods-package}
\alias{InterpretationMethods-package}
\alias{InterpretationMethods}
\title{InterpretationMethods: Facilitates the implementation of interpretation methods for social science research}
\description{
Facilitates the implementation of interpretation methods for social science researchers. The packages uses functions for interpretability methods from the packages 'iml' and 'permimp'.
}
\author{
Mirka Henninger
}
