% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_ice_ale.R
\name{pdp_ice_ale}
\alias{pdp_ice_ale}
\title{Plot PD, ICE, and ALE curves}
\usage{
pdp_ice_ale(
  pred,
  features,
  method,
  xlabel = features,
  ylabel = method,
  title = "",
  center = FALSE,
  limits = c(NA, NA),
  nCol = NA,
  alpha = 0.5,
  sample_prop = 0.5
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{features}{A character vector containing the names of the features for which the plot should be created}

\item{method}{A character string indicating the method to be applied: either "pdp", "pdp+ice", "ice", or "ale"}

\item{xlabel}{An optional character vector indicating x-axis label}

\item{ylabel}{An optional character string indicating y-axis label}

\item{title}{An optional character string indicating the title of the plot}

\item{center}{Logical indicating whether ICE curves should be centered at the minimum. Default is FALSE}

\item{limits}{An optional two-entry vector indicating the limits of the y-axis}

\item{nCol}{An optional numeric entry indicating the number of columns when plots are created for several features}

\item{alpha}{An optional numeric entry indicating the alpha-level of ICE curves}

\item{sample_prop}{An optional numeric entry between 0 and 1 indicating the percentage of individiuals that should be sampled for ICE curves}
}
\value{
a plot of type ggplotify
}
\description{
This function allows to create PD, ICE, ALE plots
ICE plots can be centered at the minimum
Plots for several features of one model can be created at the same time
}
\examples{
\dontrun{
N <- 1000
x1 <- runif(N, -1, 1)
x2 <- runif(N, -1, 1)
y <- 5 + 2 * x1 + x2 + rnorm(N,1)
dat <- data.frame(x1,x2,y)
rfmod <- randomForest::randomForest(y~., dat)
pred <- iml::Predictor$new(rfmod)
pdp <- pdp_ice_ale(pred,  "x1", method = "pdp")
ice <- pdp_ice_ale(pred,  "x1", method = "ice")
pdp_ice <- pdp_ice_ale(pred,  "x1", method = "pdp+ice")
ale <- pdp_ice_ale(pred,  c("x1", "x2"), method = "ale", ylabel = "ALE")
}

}
