% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_importance.R
\name{permutation_importance}
\alias{permutation_importance}
\title{Plot permutation importance}
\usage{
permutation_importance(
  object,
  type,
  title = "",
  xlabel = "",
  ylabel = "Permutation Importance",
  limits = c(NA, NA)
)
}
\arguments{
\item{object}{A prediction object from package iml for model-based permutation importance or an object as returned by randomForest or cforest}

\item{type}{A character indicating the type of variable importance: model, randomforest, or conditional}

\item{title}{An optional character string indicating the title of the plot}

\item{xlabel}{An optional character string indicating the x-axis label}

\item{ylabel}{An optional character string indicating the y-axis label}

\item{limits}{An optional two-entry vector indicating the limits of the y-axis}
}
\value{
a plot of type ggplot
}
\description{
This function allows to plot permutation importance based on mean squared error.
Three types are implemented: model-based permutation importance, random forest permutation importnance or conditional random forest variable importance
}
