% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_pdp_ale.R
\name{twoD_pdp_ale}
\alias{twoD_pdp_ale}
\title{Plot two-dimensional pdp, ice, ale}
\usage{
twoD_pdp_ale(
  pred,
  features,
  method,
  title = "",
  xlabel = features[1],
  ylabel = features[2],
  limits = c(NA, NA),
  show_data = FALSE,
  legend_position = "right",
  rugs = TRUE
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{features}{A character vector containing the names of two features for which the plot should be created}

\item{method}{A character string indicating the method to be applied: either "pdp", "pdp+ice", "ice", or "ale"}

\item{title}{An optional character string indicating the title of the plot}

\item{xlabel}{An optional character string indicating x-axis label (same for all panels)}

\item{ylabel}{An optional character string indicating y-axis label (same for all panels)}

\item{limits}{An optional two-entry vector indicating the limits of the y-axis}

\item{show_data}{Logical indicating whether datapoints should be plotted. Default is FALSE}

\item{legend_position}{Logical indicating whether the legend should be shown. Default is TRUE}

\item{rugs}{Logical indicating whether rugs should be shown. Default is TRUE}
}
\value{
a plot of type ggplotify
}
\description{
This function allows to plot 2-dimensional pdp, ale figures.
Please note ale for two categorical features is not yet implemented, and that the categorical feature
must be placed on the x-axis.
}
\examples{
\dontrun{
N <- 1000
x1 <- runif(N, -1, 1)
x2 <- runif(N, -1, 1)
y <- 5 + 5 * x1 * x2 + rnorm(N,1)
dat <- data.frame(x1,x2,y)
rfmod <- randomForest::randomForest(y~., dat)
pred <- iml::Predictor$new(rfmod)
(twoD_pdPlot <- twoD_pdp_ale(pred, features = c("x1", "x2"), method = "pdp"))
(twoD_alePlot <- twoD_pdp_ale(pred, features = c("x1", "x2"), method = "ale"))
}

}
