% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colored_ice.R
\name{colored_ice}
\alias{colored_ice}
\title{Plot colored ICE curves}
\usage{
colored_ice(
  pred,
  features,
  covar,
  title = "",
  xlabel = features,
  center = FALSE,
  nCol = NA,
  legend_title = covar,
  legend_position = "right"
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{features}{A character vector containing the names of the features for which the plot should be created}

\item{covar}{A character string indicating the covariate after which the ICE curves should be colored}

\item{title}{An optional character string indicating the title of the plot}

\item{xlabel}{An optional character string of the same length as the number of features indicating the x-axis label of the single plots}

\item{center}{Logical indicating whether ICE curves should be centered at the minimum. Default is FALSE}

\item{nCol}{An optional numeric entry indicating the number of columns when plots are created for several features}

\item{legend_title}{A character indicating the legend title. Default is the name of 'covar'}

\item{legend_position}{Logical indicating whether the legend should be shown. Default is TRUE}
}
\value{
a plot of type ggplotify
}
\description{
This function allows to plot colored ice predictions for a set of features, colored by a covariate
}
