% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_statistic.R
\name{interaction_statistic}
\alias{interaction_statistic}
\title{Plot feature interaction}
\usage{
interaction_statistic(
  pred,
  type = "overall",
  feature = NA,
  limits = c(0, 1),
  title = "",
  xlabel = "",
  ylabel = "Interaction Statistic"
)
}
\arguments{
\item{pred}{A prediction object from package iml}

\item{type}{A character string indicating the type of the interaction statistic. Either "overall" (Default) or "2way"}

\item{feature}{A character string indicating the name of the feature that is the basis for 2-way interactions}

\item{limits}{An optional two-entry vector indicating the y-axis limits. Default is c(0,1)}

\item{title}{An optional character string indicating the title of the plot}

\item{xlabel}{An optional character string indicating x-axis label}

\item{ylabel}{An optional character string indicating y-axis label}
}
\value{
a plot of type ggplot
}
\description{
This function allows to plot the Hamilton feature interaction statistic
}
\examples{
\dontrun{
N <- 1000
x1 <- runif(N, -1, 1)
x2 <- runif(N, -1, 1)
x3 <- runif(N, -1, 1)
y <- 5 + 5 * x1 * x2 + x3 + rnorm(N,1)
dat <- data.frame(x1,x2,x3,y)
rfmod <- randomForest::randomForest(y~., dat)
pred <- iml::Predictor$new(rfmod)
interaction_statistic(pred)
interaction_statistic(pred, type = "2way", feature = "x1")
}

}
